perf: opt gizmo ec structure
