+++ plugin-packages/editor-gizmo/demo/src/gizmo-wireframe.ts
@@ -1,9 +1,9 @@
 import type { JSONValue } from '@galacean/effects';
 import { Player } from '@galacean/effects';
-import type { GizmoVFXItem } from '@galacean/effects-plugin-editor-gizmo';
 import '@galacean/effects-plugin-editor-gizmo';
 import '@galacean/effects-plugin-model';
 import test_scene from './json/wireframe-mode.json';
+import { GizmoComponent } from 'plugin-packages/editor-gizmo/src/gizmo-component';
 
 (async () => {
   const player = new Player({
@@ -14,10 +14,10 @@ import test_scene from './json/wireframe-mode.json';
     onItemClicked: e => {
       const { player, id } = e;
       const composition = player.getCompositions()[0];
-      const item = composition.items.find(item => item.id === String(id)) as GizmoVFXItem;
+      const item = composition.items.find(item => item.id === String(id))!;
 
       console.info('itemId: ' + item.id);
-      console.info('hitBoundingKey: ' + item.hitBounding?.key);
+      console.info('hitBoundingKey: ' + item.getComponent(GizmoComponent)?.hitBounding?.key);
     },
   });
