+++ plugin-packages/editor-gizmo/demo/src/scene.ts
@@ -1,6 +1,6 @@
 import { Player } from '@galacean/effects';
-import type { GizmoVFXItem } from '@galacean/effects-plugin-editor-gizmo';
 import { primaryJSON, sceneJSON } from './assets';
+import { GizmoComponent } from 'plugin-packages/editor-gizmo/src/gizmo-component';
 
 (async () => {
   const player = new Player({
@@ -10,10 +10,10 @@ import { primaryJSON, sceneJSON } from './assets';
     onItemClicked: e => {
       const { player, id } = e;
       const composition = player.getCompositions()[0];
-      const item = composition.items.find(item => item.id === String(id)) as GizmoVFXItem;
+      const item = composition.items.find(item => item.id === String(id))!;
 
       console.info('itemId: ' + item.id);
-      console.info('hitBoundingKey: ' + item.hitBounding?.key);
+      console.info('hitBoundingKey: ' + item.getComponent(GizmoComponent)?.hitBounding?.key);
     },
   });
   const json = JSON.parse(primaryJSON);
