+++ plugin-packages/editor-gizmo/src/index.ts
@@ -1,17 +1,16 @@
-import { logger, registerPlugin } from '@galacean/effects';
+import type { VFXItemContent } from '@galacean/effects';
+import { VFXItem, logger, registerPlugin } from '@galacean/effects';
 import { EditorGizmoPlugin } from './gizmo-loader';
-import { GizmoVFXItem } from './gizmo-vfx-item';
 import { GizmoSubType } from './define';
 import { GeometryType, createGeometry } from './geometry';
 
-registerPlugin('editor-gizmo', EditorGizmoPlugin, GizmoVFXItem);
+registerPlugin('editor-gizmo', EditorGizmoPlugin, VFXItem<VFXItemContent>);
 
 export { DirectionLightData } from './geometry/direction-light';
 export {
   GizmoSubType,
   GeometryType,
   createGeometry,
-  GizmoVFXItem,
 };
 
 export const version = __VERSION__;
